fix: improve screen layout
